Create Cricket_Manager_Version_2(SHELVED)
This project has been shelved for now. Here onwards, I will instead be using the comfort of the Unity engine to develop further games in C#. The whole nature of the game however has become a tad boring to me as this is the second version of it, and I know I can develop it, and I don't see many challenges besides ensuring the game is not hardcoded as it is in this document. Nonetheless, I know how to clean it up and ensure it is not so hard-coded, so here again, the challenge is more or less gone. As such, I will be moving on to new projects!